requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-js-asset==3.0.1
django-mptt==0.16.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-silk==5.3.2
django-tag-parser==3.2
djangorestframework==3.15.2
gprof2dot==2024.6.6
pillow==11.0.0
pycodestyle==2.12.1
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
